requirements.txt
asgiref==3.4.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.6
django-environ==0.8.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
filelock==3.4.2
gunicorn==20.1.0
mysqlclient==2.1.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
tzdata==2021.5
virtualenv==20.13.0
whitenoise==5.3.0
